move install state builder to manager
