+++ test/uplink/packages/install/execute_test.exs
@@ -1,0 +1,3 @@
+defmodule Uplink.Packages.Install.ExecuteTest do
+  use ExUnit.Case
+end
